batch writing working

+++ Froodz/Model/Group.swift
@@ -15,7 +15,7 @@ struct Group : Codable {
     let documentId: String
     let groupName: String
     let code : String
-    let users : [String: Int]
+    let users : [String: Double]
     
     enum CodingKeys: String, CodingKey {
         case documentId
